requirements.txt
APScheduler==3.10.4
certifi==2024.7.4
charset-normalizer==3.3.2
et-xmlfile==1.1.0
greenlet==3.0.3
idna==3.7
load-dotenv==0.1.0
numpy==2.0.0
openpyxl==3.1.5
pandas==2.2.2
psycopg2-binary==2.9.9
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
workbook==1.1
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
schedule
